requirements.txt
Flask
pymongo
mysql-connector-python
redis
gunicorn
Werkzeug
python-dotenv
